minor refactoring and renamings
